refactor: split Chart/cairo out of the library into a plot internal library
Chart, Chart-cairo and colour were declared at package level, so the
library inherited them, and the library is a single stanza. Any
consumer therefore compiled the whole tree plus cairo and pango to
reach pure numeric code.

Measured before this change: 19 of 62 library modules imported
Graphics.Rendering.Chart directly, and 49 of 62 were transitively
tainted -- only 13 were clean.

src/ is now the pure numeric core (deps: base, mwc-random, vector).
plot/ is a new INTERNAL library holding every layout, depending on the
core plus Chart/Chart-cairo/colour. Internal rather than public
because no external consumer needs plotting, and internal sublibraries
are invisible to build-depends outside the package -- a deterministic
guarantee rather than a cabal flag the solver could flip.

Moving the dep lines to the executables stanza, as first proposed,
would not have compiled: plotting was not confined to leaf modules
(TickPath exported tickPathLayout :: TickPath -> Layout Double Double
beside the path math, and Panoptic.NId reached cairo transitively
through Pricing.PriceDeformation).

Behaviour-preserving: all 49 committed PNGs in outputs/ regenerate
byte-identically from the rebuilt exe.
