+++ plot/Plot/Payoffs/RangeAccrualNote.hs
@@ -1,30 +1,24 @@
-{-# LANGUAGE PatternSynonyms #-}
-
-module Payoffs.RangeAccrualNote
-  ( payoff
-  , rangeAccrualNote
+-- | Chart layouts for 'Payoffs.RangeAccrualNote'.  Split out of the pure module so that
+-- consumers of the numeric core do not compile Chart/cairo.
+module Plot.Payoffs.RangeAccrualNote
+  ( plotPayoff
   ) where
 
 import qualified Payoffs.Payoff as Payoff
-
 import SqrtGrid
   ( SqrtPriceX96(..)
   , PayoffX96(..)
   , SqrtPlot
   )
-
+import Plotting.PlotSqrt (PlotY(..), plotSqrtFunction)
 import StrikeX96 (StrikeX96(..))
 import OptionRatio (OptionRatio(..))
+import Payoffs.RangeAccrualNote
 
-payoff :: SqrtPriceX96 -> StrikeX96 -> OptionRatio -> PayoffX96
-payoff (SqrtPriceX96 p) k@(StrikeX96 kRaw) ro@(OptionRatio r)
-  | p < lowerBound k ro  = PayoffX96 0
-  | p < kRaw             = PayoffX96 $ floor (fromInteger (lowerArmNumerator p kRaw r) / (r - 1))
-  | p < upperBound k ro  = PayoffX96 $ floor (fromInteger (upperArmNumerator p kRaw r) / (r - 1))
-  | otherwise            = PayoffX96 0
 
-rangeAccrualNote :: StrikeX96 -> OptionRatio -> Payoff.Payoff SqrtPriceX96
-rangeAccrualNote k r = Payoff.Payoff (\p -> payoff p k r)
+plotPayoff :: FilePath -> SqrtPlot -> StrikeX96 -> OptionRatio -> IO ()
+plotPayoff path config k r =
+  plotSqrtFunction path config PayoffY [Payoff.runPayoff (rangeAccrualNote k r)]
 
 -- Lower boundary: kappa / sqrt(r)  in X96
 lowerBound :: StrikeX96 -> OptionRatio -> Integer
