+++ src/Payoffs/CLMMPosition.hs
@@ -8,11 +8,8 @@ module Payoffs.CLMMPosition
   , fromCall
   , fromPut
   , chunkFromStrike
-  , scaledVsUnitLayout
+  , strikeAndRatio
   , toPayoff
-  , plotPayoff
-  , clmmEtaLayout
-  , rhsPayoffLayout
   , fromDelta
   ) where
 
@@ -23,8 +20,6 @@ import qualified Payoffs.CoveredCall as CC
 import qualified Payoffs.CashSecuredPut as CSP
 import qualified Payoffs.RangeAccrualNote as RAN
 
-import Graphics.Rendering.Chart.Easy (Layout)
-
 import SqrtGrid
   ( SqrtPriceX96(..)
   , PayoffX96(..)
@@ -34,7 +29,6 @@ import SqrtGrid
   , sqrtPriceX96
   , tickFromSqrtPriceX96
   )
-import Plotting.PlotSqrt (PlotY(..), plotSqrtFunction, sqrtFunctionLayout)
 
 import Pricing.PriceDeformation
   ( EtaX96
@@ -131,66 +125,3 @@ fromDelta spot r d =
 
 toPayoff :: CLMMPosition -> Payoff.Payoff SqrtPriceX96
 toPayoff = clmmPayoff
-
-plotPayoff :: FilePath -> SqrtPlot -> StrikeX96 -> OptionRatio -> IO ()
-plotPayoff path config k r =
-  plotSqrtFunction path config PayoffY
-    [ Payoff.runPayoff (toPayoff (fromCall k r))
-    , Payoff.runPayoff (toPayoff (fromPut  k r))
-    ]
-
--- x = undeformed p_{1/2}(i); y = π(p_{1/2}(i; η))
-payoffAtEta
-  :: EtaX96
-  -> Payoff.Payoff SqrtPriceX96
-  -> SqrtPriceX96
-  -> PayoffX96
-payoffAtEta eta payoff sample =
-  case deformedSqrtPriceX96 eta (tickFromSqrtPriceX96 sample) of
-    Nothing -> PayoffX96 0
-    Just deformed -> Payoff.runPayoff payoff deformed
-
-clmmEtaLayout
-  :: SqrtPlot
-  -> StrikeX96
-  -> OptionRatio
-  -> [(String, EtaX96)]
-  -> Layout Double Double
-clmmEtaLayout config k r labeledEtas =
-  let
-    position = toPayoff (fromCall k r)
-  in
-    sqrtFunctionLayout config PayoffY
-      [ (label, payoffAtEta eta position)
-      | (label, eta) <- labeledEtas
-      ]
-
-rhsPayoffLayout
-  :: SqrtPlot
-  -> StrikeX96
-  -> OptionRatio
-  -> EtaX96
-  -> Layout Double Double
-rhsPayoffLayout config k r warpedEta =
-  let
-    clmm = toPayoff (fromCall k r)
-  in
-    sqrtFunctionLayout config PayoffY
-      [ ("Covered Call", payoffAtEta BASE_ETA (CC.coveredCall k))
-      , ("Range Accrual", payoffAtEta BASE_ETA (RAN.rangeAccrualNote k r))
-      , ("CLMM η = 1/2", payoffAtEta BASE_ETA clmm)
-      , ("CLMM η = 2/3", payoffAtEta warpedEta clmm)
-      ]
-
--- | Scaled (chunk principal) vs unit (per token0 notional) on one sqrt axis.
-scaledVsUnitLayout
-  :: SqrtPlot
-  -> LiquidityChunk
-  -> Layout Double Double
-scaledVsUnitLayout config ch =
-  let (k, r) = strikeAndRatio ch
-      unit   = fromCall k r
-  in  sqrtFunctionLayout config PayoffY
-        [ ("CLMM unit (amount0 = 1 token0)", Payoff.runPayoff (toPayoff unit))
-        , ("CLMM × amount0 (chunk principal)", Payoff.runPayoff (toPayoff (fromChunk ch)))
-        ]
