+++ src/Payoffs/CoveredCall.hs
@@ -2,7 +2,6 @@ module Payoffs.CoveredCall
   ( payoff
   , coveredCall
   , strikeDerivative
-  , plotPayoff
   ) where
 
 import qualified Payoffs.Payoff as Payoff
@@ -13,8 +12,6 @@ import SqrtGrid
   , PayoffX96(..)
   , SqrtPlot
   )
-import Plotting.PlotSqrt (PlotY(..), plotSqrtFunction)
-
 
 payoff
   :: SqrtPriceX96
@@ -30,13 +27,11 @@ payoff
       PayoffX96 $
         p - max (p - k) 0
 
-
 coveredCall
   :: Strike.StrikeX96
  -> Payoff.Payoff SqrtPriceX96
 coveredCall strikePrice =
   Payoff.Payoff (`payoff` strikePrice)
-
 
 strikeDerivative
   :: Strike.StrikeX96
@@ -50,32 +45,3 @@ strikeDerivative
       Strike.StrikeSlope 1
   | otherwise =
       Strike.StrikeSlope 0
-
-
-plotPayoff
-  :: FilePath
-  -> SqrtPlot
-  -> Strike.StrikeX96
-  -> Strike.StrikeVariation
-  -> IO ()
-plotPayoff path config strikePrice variation =
-  let
-    originalPayoff =
-      coveredCall strikePrice
-
-    variedStrike =
-      Strike.applyStrikeVariation
-        strikePrice
-        variation
-
-    variedPayoff =
-      coveredCall variedStrike
-
-  in
-    plotSqrtFunction
-      path
-      config
-      PayoffY
-      [ Payoff.runPayoff originalPayoff
-      , Payoff.runPayoff variedPayoff
-      ]
