+++ src/Payoffs/LadderPosition.hs
@@ -20,11 +20,7 @@ module Payoffs.LadderPosition
   , ladderReturnQ96
   , logPortfolioQ96
   , cOfS
-  , ladderLayout
-  , ladderDensityLayout
   ) where
-
-import Graphics.Rendering.Chart.Easy (Layout)
 
 import Liquidity.LiquidityChunk
   ( LiquidityChunk
@@ -48,7 +44,6 @@ import Liquidity.LiquidityGrid
 import qualified Payoffs.CLMMPosition as CLMM
 import Payoffs.Log (logPortfolioQ96)
 import qualified Payoffs.Payoff as Payoff
-import Plotting.PlotSqrt (PlotY(..), sqrtFunctionLayout)
 import SqrtGrid
   ( PayoffX96(..)
   , SqrtPlot
@@ -153,28 +148,3 @@ cOfS s =
   let lnLam = log 1.0001
       sD = fromIntegral s
   in  1 / (2 * (lnLam * sD / 4 + (1 - 1.0001 ** (negate sD / 2)) / 2))
-
--- | T1/N_1 vs c(S)·logPortfolio on one sqrt axis (Theorem 10 overlay).
-ladderLayout :: SqrtPlot -> Ladder -> Layout Double Double
-ladderLayout config l =
-  let s = ladderHi l - ladderLo l
-      c = cOfS s
-      pStar = sqrtPriceX96 (ladderStar l)
-      scaledT0 p = let PayoffX96 y = logPortfolioQ96 p pStar in PayoffX96 (floor (c * fromIntegral y))
-  in  sqrtFunctionLayout config PayoffY
-        [ ("T1/N_1 (ladder, ξ*)", ladderReturnQ96 l)
-        , ("c(S)·logPortfolio (T0)", scaledT0)
-        ]
-
--- | Rung liquidities L(i_x) as a step function of sqrt price.
-ladderDensityLayout :: SqrtPlot -> Ladder -> Layout Double Double
-ladderDensityLayout config l =
-  let chs = ladderChunks l
-      d = unTickSpacing (ladderSpacing l)
-      at (SqrtPriceX96 p) =
-        PayoffX96 $ sum [ chunkLiquidity ch
-                        | ch <- chs
-                        , let SqrtPriceX96 a = sqrtPriceX96 (chunkTickLower ch)
-                        , let SqrtPriceX96 b = sqrtPriceX96 (chunkTickLower ch + d)
-                        , a <= p && p < b ]
-  in  sqrtFunctionLayout config PayoffY [ ("L(i_x) = ΔQ·ℓ(ξ*,ι;x)", at) ]
