+++ src/Payoffs/LvrRate.hs
@@ -28,16 +28,13 @@ module Payoffs.LvrRate
   , lvrRate
   , lvrRateOn
   , lvrRateTable
-  , lvrRateLayout
   ) where
-
-import Graphics.Rendering.Chart.Easy (Layout)
 
 import Liquidity.LiquidityChunk (LiquidityChunk, chunkAmount0)
 import Panoptic.LegChunk (legChunks)
 import Panoptic.MintPlan (MintPlan)
 import Payoffs.HolderPath (Regime(..), composedPath)
-import Payoffs.PathAccrual (Accrual(..), Step(..), Tag(..), addAccrual, linesLayout, pathAccrual, pattern PIPS_ONE, stepVolume1, zeroAccrual)
+import Payoffs.PathAccrual (Accrual(..), Step(..), Tag(..), addAccrual, pathAccrual, pattern PIPS_ONE, stepVolume1, zeroAccrual)
 import Pricing.Stremia (FeePips, unFeePips)
 import SqrtGrid (PayoffX96(..), Tick, mulDiv)
 
@@ -68,14 +65,3 @@ lvrRateOn seed i0 n transAmp phi band chs s =
 
 lvrRateTable :: Int -> Tick -> Int -> Int -> FeePips -> Int -> MintPlan -> [Int] -> [(Integer, Integer)]
 lvrRateTable seed i0 n transAmp phi band plan ss = [ (toInteger s, lvrRate seed i0 n transAmp phi band plan s) | s <- ss ]
-
--- | λ vs s for several φ, naive band (crosses zero at 2φ ticks) and rational band (≥ 0).
--- Axes: ticks, pips.
-lvrRateLayout :: Int -> Tick -> Int -> Int -> MintPlan -> [FeePips] -> [Int] -> Layout Double Double
-lvrRateLayout seed i0 n transAmp plan phis ss =
-  linesLayout "λ_{X/M}: LVR_net per unit arb volume vs external step s (holder inactive)"
-    "external step s (ticks / round)" "LVR_net / Σ amount_in (pips)"
-    (concat
-      [ [ ("φ = " ++ show (unFeePips phi) ++ " pips, band φ (naive)", lvrRateTable seed i0 n transAmp phi (naiveBandTicks phi) plan ss)
-        , ("φ = " ++ show (unFeePips phi) ++ " pips, band 2φ (rational)", lvrRateTable seed i0 n transAmp phi (rationalBandTicks phi) plan ss) ]
-      | phi <- phis ])
