+++ src/Payoffs/ReplicaDelta.hs
@@ -14,12 +14,9 @@ module Payoffs.ReplicaDelta
   ( replicaDelta
   , legDelta
   , principalDelta
-  , replicaDeltaLayout
   ) where
 
-import Graphics.Rendering.Chart.Easy (Layout)
-
-import Greeks.Delta (PayoffDelta(..), PriceDeltaX96(..), deltaOfPayoff, payoffDeltaLayout)
+import Greeks.Delta (PayoffDelta(..), PriceDeltaX96(..), deltaOfPayoff)
 import Liquidity.LiquidityChunk (LiquidityChunk, chunkAmount0, chunkLiquidity, chunkTickLower, chunkTickUpper)
 import Panoptic.LegChunk (legChunk)
 import Panoptic.MintPlan (MintPlan(..), fourLegNumLegs)
@@ -53,11 +50,3 @@ replicaDelta plan =
     PriceDeltaX96 $ sum
       [ d | leg <- [0 .. fourLegNumLegs (mintTokenId plan) - 1]
           , let PriceDeltaX96 d = legDelta plan leg p ]
-
--- | Closed form vs the generic finite-difference instance on the same axis.
-replicaDeltaLayout :: SqrtPlot -> MintPlan -> SqrtPriceX96 -> Layout Double Double
-replicaDeltaLayout config plan pStar =
-  payoffDeltaLayout config
-    [ ("Δ̂^σ closed form (Σ_leg)", replicaDelta plan)
-    , ("∂_P π̂^σ central difference", deltaOfPayoff (fourLegReplica plan pStar))
-    ]
