+++ src/Pricing/AdaptiveStremia.hs
@@ -21,12 +21,8 @@ module Pricing.AdaptiveStremia
   , sigmoid
   , expXg4
   , pathVolatility
-  , adaptiveFeeLayout
   ) where
 
-import Graphics.Rendering.Chart.Easy (Layout)
-
-import Payoffs.PathAccrual (linesLayout)
 import Pricing.Stremia (FeePips, mkFeePips, unFeePips)
 import SqrtGrid (Tick)
 
@@ -125,11 +121,3 @@ pathVolatility dt ticks
           m = s `div` n
           sq = sum [ (toInteger i - m) ^ (2 :: Int) | i <- ticks ]
       in  Volatility ((sq * dt) `div` (n * dt))
-
--- | Fee (pips) vs volatility (oracle units) for a configuration. Axes: uint88 units × pips.
-adaptiveFeeLayout :: AdaptiveStremia -> [Integer] -> Layout Double Double
-adaptiveFeeLayout c vols =
-  linesLayout "Algebra AdaptiveFee.getFee: baseFee + σ₁ + σ₂ (integer-exact port)"
-    "volatility (oracle uint88 units, before /15)" "fee (pips)"
-    [ ("getFee", [ (v, unFeePips (adaptiveFeePips c (Volatility v))) | v <- vols ])
-    , ("baseFee + α₁ + α₂ cap", [ (v, baseFee c + alpha1 c + alpha2 c) | v <- vols ]) ]
